// server/src/api/types.ts
export interface AuthRequired {
  Authorization: string;
}
// ...
export interface ListApartmentsRequest {
  filter?: ListApartmentsFilter;
  maxPerPage?: number;
  page?: number;
}

export interface ListApartmentsFilter {
  realtorId?: string;
  rented?: boolean;
  sizeRange?: {
    min: number;
    max: number;
  };
  priceRange?: {
    min: number;
    max: number;
  };
  numberOfRooms?: {
    min: number;
    max: number;
  };
  viewport?: Viewport;
}

export interface ListApartmentsResponse {
  results: ApartmentDetails[];
  totalResults: number;
  pageCount: number;
}

export interface ApartmentDetails {
  apartmentId: string;
  info: ApartmentInfo;
  realtor: Realtor;
  dateAdded: Date;
}

export interface ApartmentInfo {
  floorArea: number;
  pricePerMonth: number;
  numberOfRooms: number;
  coordinates: Coordinates;
  rented: boolean;
}

export interface Viewport {
  southWest: Coordinates;
  northEast: Coordinates;
}
// ...
export type ListApartments_Response = {
  kind: "success";
  data: ListApartmentsResponse;
} | {
  kind: "unauthorized";
  data: string;
};